feat: add service forgot-password, Update commission rate to 25%

+++ src/main/java/com/sa/event_mng/model/projection/EventRevenueStatsAdminProjection.java
@@ -1,5 +1,7 @@
 package com.sa.event_mng.model.projection;
 
+import java.math.BigDecimal;
+
 public interface EventRevenueStatsAdminProjection {
-    Long getTotalRevenue();
+    BigDecimal getTotalRevenue();
 }
